Refactor: seed 합치기
