--- mock/companyIds.ts
@@ -1,52 +1,0 @@
-export const companyIds = [
-  "003f12ac-a3f1-4f00-a441-8fd719793a8e",
-  "07e4c45d-d143-4b35-8886-e95c41540de9",
-  "0d366a8b-a929-4ca3-ad0f-1262d475b5b2",
-  "0ebfcad9-1c0d-4ddd-a361-f4b60085b8a1",
-  "17087a8b-f736-48f3-bdf5-9609d9d3c8f6",
-  "17f2022a-ec01-4efe-8e73-4587d6e3521d",
-  "1ba738f4-4b1b-4190-8998-9bb851b12110",
-  "1bd7434e-62ac-40a0-be20-324834146f56",
-  "1da2e85c-2128-4d59-8f10-b1a17a308ac1",
-  "2502f28a-b46e-4b58-b4c2-4066c6ebe889",
-  "25cf3d75-d4f5-49e4-9cd0-8c19f4463318",
-  "2651e520-08bb-496c-8ba6-d74d4f1f578c",
-  "2afa9bf2-6c05-4bae-84dc-77dfde2c18a8",
-  "36489035-9fa8-4450-8aa7-ba7af1b3b350",
-  "37b9170a-6095-447b-a9e9-bc0d009fe2f5",
-  "3a12c13c-e200-470f-a9eb-86630629ae75",
-  "3e2d5934-795b-46f1-a7e8-321078e07dae",
-  "45cd3b54-0267-449f-b9a8-14da408a3ed1",
-  "4f711337-c769-4d12-a466-62fd99832aea",
-  "5187106c-a7fc-4aca-a44c-1a0cb15d46fc",
-  "5427a048-8d14-486f-8d5a-76fa9bfbdf2b",
-  "614bb390-18a8-4527-93f7-e7dc09208273",
-  "67f0885d-30ba-4b69-bb12-7ac209ec74e0",
-  "689b2557-93c8-4ecb-b734-62eda4114cee",
-  "6dc2e52d-0a93-4983-a68b-f7937f35bf46",
-  "72e0f6ee-69e4-4793-b879-a91793d2fd0c",
-  "757e158e-b201-445a-8a66-87e9ce4a3452",
-  "76499738-67ad-4852-8978-16f4eda3376b",
-  "7f321853-8f4a-4ef8-a60b-55fa61e2afcf",
-  "888021ae-3340-4f53-b6c0-98706015ff56",
-  "8daed5cf-b229-4d50-8682-4d024f69d75f",
-  "8fb59eb0-809a-4cf3-b14c-cc5f3f5cbcb5",
-  "93d188dd-1255-44c4-83ae-78af82ec8c4f",
-  "96550cfe-d2c9-40a4-bf78-64c07de79686",
-  "9c4d66a1-3486-48eb-8f8e-02d8a360bd30",
-  "9cb726a9-7f4c-40aa-9ab7-ea263d196137",
-  "9d0e0bdf-284d-4472-b84d-22d6458df96c",
-  "9fa0e0ba-cb3c-492b-b251-ae8c952b7a87",
-  "acbc0a4c-bba5-49de-880a-90649e8a2340",
-  "adef3d6e-d43e-4667-afaa-da2b9dd89ed9",
-  "b0620173-242a-4910-8e90-8b76dcfaca32",
-  "bf163ac3-9f8b-431f-86bc-dc80e4220efb",
-  "c92bb21d-92d0-4480-ad6d-773da336b066",
-  "e0fa5ea5-ce1b-47a0-9cb0-e61612ce40f6",
-  "e36195db-b660-438b-9c84-41481419f710",
-  "e4717778-b585-4695-ae7b-8e310a9744c0",
-  "e85a6e96-a5ff-40a7-8807-2781161f3f82",
-  "e85fc99e-8adc-485c-b6ad-35dd4367be09",
-  "f61583ac-dfcc-4083-a19a-6b37588d5020",
-  "fd57e70f-b19c-4d88-a798-bad6960abdfd",
-];
